// packages/typespec-ts/test/modularIntegration/generated/lro/modular/src/api/models.ts
// Copyright (c) Microsoft Corporation.
// Licensed under the MIT license.

/** Details about a user. */
export interface User {
  /** The name of user. */
  readonly name: string;
  /** The role of user */
  role: string;
}

/** */
export interface User {
  /** The name of user. */
  readonly name: string;
  /** The role of user */
  role: string;
}

/** */
export interface User {
  /** The name of user. */
  readonly name: string;
  /** The role of user */
  role: string;
}